+++ CHANGELOG.md
@@ -1,4 +1,3 @@
 # LibSFDropDown
 
-* Fixed hiding the color swatch button.
-* Now the iconOnly key can be used in info.list.
+* Fixed lib methods error.